requirements.txt
# 核心依賴 - Discord Bot
discord.py==2.6.0
python-dotenv==1.1.1

# 核心依賴 - Playwright 爬蟲
playwright==1.58.0

# 核心依賴 - 圖像處理與 OCR
Pillow==11.3.0
opencv-python==4.11.0.86
numpy==2.2.6
ddddocr==1.6.0

# 核心依賴 - 檔案鎖定（並發控制）
filelock==3.24.3

# 輔助依賴 - class_id.py 使用（可選）
requests==2.32.5
